client, packet: remove RemoteAddr
There should be no reason to send replies specifically to the remote
address as denoted in the ethernet frame. ARP packets contain a sender
address, to which we can direct our replies. Nothing in the spec
suggests that the ethernet frame could or should overrule the address in
the ARP request. On the contrary, the opposite might be the case.