client: make WriteTo use addr argument for the Ethernet frame
According to RFC 826, the target hardware address in the ARP packet
isn't important (in one place even marked as "don't care"), and it
specifically says that the address lacks meaning in the request form.
Nevertheless, hardware exists that (regardless of whether or not it's
compliant) imposes requirements on the target address.

Previously, the WriteTo function in this library has disregarded its
addr parameter and instead used the target MAC address from inside the
ARP packet for the Ethernet destination address. This made it
impossible to communicate with hardware that has expectations on the
target MAC address field as you couldn't use an ARP target MAC address
that's distinct from the Ethernet frame's destination address.

fixes #20